Reword positioning of multi-targeting
I don't think we should position multi-targeting as an alternative to .NET
Standard. While it could be used this way, that's not what we recommend
because it forwards the multi-targeting requirement to the consumer. Rather,
our recommendation is to build bait-and-switch-style .NET Standard libraries
where you always target .NET Standard while also providing framework-specific
implementations.